+++ src/dose/dose_internal.ss/src/LeveledLockHelper.cpp
@@ -154,7 +154,7 @@ namespace Internal
         }
         else
         {
-            return mapIt->second.size();
+            return static_cast<int>(mapIt->second.size());
         }
     }
 }
